[WIP] SchemaV16: Implement Decks
This converts decks.py to DecksV16.kt as it was in Anki 2.1.34

To move to the v16 schema, we need to upgrade the classes
associated with the collection columns

These have moved from JSON in a single column to one row per protobuf

After this is merged, we can start migrating Decks.java to use the
same interface and consolidate functionality.

Once the interfaces are the same, we can build a new "Collection" class
with the new table-based classes, flip to opening the collection via
the v16 upgrade mechanism, and then check for bugs.

Some of this functionality is explicitly WIP (DecksBackend.kt)

Related: 8988