lint: remove "debug" config
After moving to AGP 7, we can no longer apply lintConfig on
`task.doFirst`

Since we're not using the baseline/debug config, we might as well
simplify this file and remove the implementation.

Baseline usage was never properly implemented, and was on Travis CI
which we are no longer using.

Related: 9338