lint: fix UnspecifiedImmutableFlag
Lint was added in AGP 7
http://googlesamples.github.io/android-custom-lint-rules/user-guide.html

Apps targeting Android 12 and higher must specify either
`FLAG_IMMUTABLE` or `FLAG_MUTABLE` when constructing a `PendingIntent`.

Since we don't have `FLAG_IMMUTABLE` (API 23), we define this in Compat
and pass it in with a new compat method

We use `FLAG_IMMUTABLE`.
`FLAG_MUTABLE` should only be used when some functionality relies on
modifying the underlying intent, e.g. any PendingIntent that needs
to be used with inline reply or bubbles.

https://developer.android.com/reference/android/app/PendingIntent#FLAG_IMMUTABLE
https://developer.android.com/reference/android/app/PendingIntent#FLAG_MUTABLE

Related: Upgrade to AGP 7: 9338