feat: obtain optional conf using get_config
Anki intends to move to this new interface

https://github.com/ankitects/anki/blob/187944615e24e2e052c7b1a9062f12618b00ac1e/pylib/anki/config.py#L4-L18

This is more complex than appears: Anki uses a method returning an Any
This is actually any one of the JSON types. We can't do this here
so we add separate methods. This is possible as the types in the
"default" are different.

Related: 8988