fix(deck-options) - Sync Graduating Intervals
We were sending a two-element array, rather than a three element
array. This was incorrect, but the third element was unused, so it
was fine.

The new Anki Desktop backend serialised this array differently, which
broke syncing the graduating/easy interval

We fix this by adding a third unused element to dconf.new.ints

We default to 7, as this is what Anki Desktop used to use

Checked with dae and he's not fussed with the value as it's currently
unused.

Fixes 8889