NF: simplify getDropDownDecks
getDropDownDecks was overloaded, and both methods did different things. One
compute the deck, the other one return the list already computed. Furthermore,
the one computing took the collection as argument, which was useless given that
it is a member of the object.

So I renamed the computing one from get to compute, and I removed its parameter