NF: remove mDeckId from DeckSpinnerSelection
selectDropDownItem was buggy. It only worked as expected if the list of deck
starts with "All decks". Otherwise (i.e. in the note editor) mDecksId has an
off-by-one error.

This is not actually important because mDeckId was used in a single place, in
getDeckId from Statistics. All other users of the deck spinner (card browser and
note editor) save the deck id themselves.

This update statistics so that it also keep track of deck id itself. This ensure
that there is no need to correct the bug as the variable is not used anymore.

The member is called mStatsDeckId so that it can't be mixed with mDeckId.