NF: Add more details to Kotlin Migration
Some details were not clear during my first try of this feature.

In particular, I didn't have the "commit" window. Instead I had a different
non-modal window, which did not offer as many option. So I had to explain how to
get the modal window.

I also added instructions about what to do if build fail, at the place where we
expect failure.

I also added sub-sections. This way, if the process fail, it's easier to know
where we are.