feat(scoped-storage): Migrate User Data context
Context for the "Migrate User Data" operation: this allows a background
task to migrate data from the AnkiDroid collection folder to a new
folder without the need for blocking the user from interacting with
AnkiDroid.

This task must be preemptable: so operations should be split into
sub-operations whenever possible. Hence having an Operation return
List<Operation>, this allows for breaking up a large "move folder"
into shorter "move file/folder" operations

Exception reporting must context-sensitive, and may desire to change
how an operation (or execution of the list of operations) is handled
based on both the operation performed, and the number of exceptions