Included both classes of Replay button
The button class now has both `replaybutton` and `replay-button` classes. The styles of the classes are defined separately so old code doesn't break. Fill of the svg is explicitly mentioned in it's html code.

NOTE: using the `night_mode` class overrides the styling done. So was not able to apply different fills for light and dark mode. But button very visible in both modes.