Add `replaybutton` css class for anki ecosystem compatibility (#10350)
* Changed SVG type and name of `replaybutton`

changed class name of button from `replaybutton` to `replay-button` to be compatible with anki desktop button styling.
changed SVG type from `<polygon>` to `<circle><path>` as defined in anki desktop to be compatible.

* Included both classes of Replay button

The button class now has both `replaybutton` and `replay-button` classes. The styles of the classes are defined separately so old code doesn't break. Fill of the svg is explicitly mentioned in it's html code.

NOTE: using the `night_mode` class overrides the styling done. So was not able to apply different fills for light and dark mode. But button very visible in both modes.

* Changed SVG type and added class for Replay Button

Added a class to include both classes `replaybutton` and `replay-button` to be compatible with anki desktop button styling as well as the old code.
Changed SVG type from `<polygon>` to `<circle><path>` as defined in Anki desktop