NF: MoveFileOrDirectoryTest consistent with other Move (#10373)
I know explicitly state that in MoveFile and MoveDirectory, the destination is
the actual new path of the file/directory. (And not the directory where the
copied element will be)

For the sake of consistency, I change MoveFileOrDirectory so that its
destination is also the new path (and not the path of the directory containing
the copied file).

I discovered it while trying to make more specific test, where the
source/destination of elements were asserted. Those test are now committed too.

Co-authored-by: David Allison <62114487+david-allison@users.noreply.github.com>