`contentOfDirectory` may throw
This was the case in #10358 due to a permission issue with newDirectoryStream.
In this case, `listFiles` returned null. In this case, throwing makes more sens,
as it indicates that something was wrong.

UI will need to catch it eventually and ask to contact AnkiDroid developer for
help, because when this occurs there is nothing the app can do anymore.