Add string ressources for scoped-storage migration.
Those resources are added before being used in order to allow a maximum number
of translators to translate them. It will also allow us to reach to translators
on all platforms we have.

Maybe we won't need all strings. For example, AnkiDroid may or may not be slowed
down during migration. We'll discover during tests. It's best to have the
strings ready just in case. I also expect that some strings will be used in
multiple messages. E.g. a link to scoped-storage explanation will probably be
used in most messages.

English is not my mother-tongue, so the goal is not to have the definitive
string here, it's to have something to start the process, this way we can
discuss how to improve our message and what string we expect we may potentially need.