Allow hiding answer buttons only (non-fullscreen)
On e-ink devices flash a lot due to the abrupt white->black change.
Enabling fullscreen mode is not sufficient as it hides the review
count which can be useful and is fairly unobtrusive.

To correctly use this mode you must use gestures or a bluetooth
controller. Add preference checks to make sure these requirements
are met before enabling.