Add option to show all deck stats by default
This option is useful to anyone who wants to know "how many cards did I
study today (in total), and how many cards are due tomorrow (in total)".
In this situation it can save a few clicks and scrolls which can be slow
on e-ink displays.

Some additional context:
If using a single "To Review" filtered deck then stats are pretty
meaningless. For instance "Studied 3 cards in 0 minutes today". To
find the real statistics after studying with a filtered deck I have
to select "All decks" and then I correctly see "Studied 491 cards in
25 minutes today".