`hasFiles` raises exception (#10398)
The primary goal how this commit is to have a clear way to catch #10358.
It changes the way the method behaves when method assumption are not satisfied;
and raising an exception seems at least as acceptable than returning null. Plus
it will lead to more specific error cause if we care about them.