NF: remove `createDirectory`
It was confusing to have both `createDirectory` and
`createTransientDirectory`. Plus, we almost never really need a `Directory` but
its underlying `File`.

This remove confusion in which method to use, what are the difference between
them, and it remove even some `.direction`, at the cost of two helpers methods
that encapsulate File in Direction when required