feat(scoped-storage): Move Conflicted File
**Rationale**

This handled the 'conflicted file' edge case when performing
a scoped storage migration:

A file `AnkiDroid/A.txt` is wanted to be moved to
`scoped directory/A.txt`, but `A.txt` already exists
in the destination AND is different.

So move it to `AnkiDroid/conflict/A.txt`

**Implementation**

An operation which accepts a rooted relative path and a source file
and moves the source file to '<root>/conflict/<path>'.

The 'conflict' directory will be ignored during the migration
 so this allows progress and completion of the migration
 even with conflicts.

This operation handles conflicts, and creation of the new directories