[Nullity] TimePreference.kt
Issue:
mTimePicker is nullable and is being used as a param while
calling the methods of Compat.java. This is working now since
Compat is in java without null contracts. However, null
annotations have to be added to Compat.java for further migration
to Kotlin(see: https://github.com/ankidroid/Anki-Android/pull/10193),
and since timePicker param is already overridden as non-null in
CompatV23.kt, same nullity must be assigned in Compat.java.

Fix: Change mTimePicker to a non-null lateinit var,
     Remove redundant non-null assertions(!!)