Migrate FilterResultsUtils.java to kotlin
This migration from java to kotlin doesn't follow the kotlin migration guide(to add one extra rename commit)
because the java code used in the class doesn't allow for a translation to kotlin. The FilterResultsUtils
class was used as a utility class to fill and provide FilterResults objects to 3 Filter implementations. However,
in the code, the objects returned by FilterResults were just filler objects to satisfy compilation.

So the solution to migrate the file was deletion of FilterResultsUtils and replacement of its usage with simple
FilterResults instantiation.