feat: Lock database
We can't lock the database via file locking on macOS under Robolectric
So instead, we lock the database in code.

This is better than locking via `CollectionHelper` as it's not opt in

So, we use `Storage`, where the Collection is created

Throws `SQLiteDatabaseLockedException` if an attempt to access the DB
is made when the database is locked.