feat: Migrate Essential Files
COPIES essential files to a scoped location. Locking the collection
to prevent corruption.

This is a fast operation, and is designed to block the foreground with
a spinner while it occurs.

This exists to move 'deckPath' quickly, so a user can continue under a
scoped location while a 'MoveUserData' executes in the background

This only copies:
* collection.anki2 + journal
* collection.media.ad.db2 + journal
* collection.log
* .nomedia