refactor: extract TreeNode from AbstractDeckTreeNode
AbstractDeckTreeNode has two responsibilities: as a base class for a
deck to be displayed, and as a tree.

The combination of the two led to the recursive structure:

```kotlin
AbstractDeckTreeNode<T : AbstractDeckTreeNode<T>?>
```

which is difficult to express in Kotlin.

As well as breaking the SRP, and being a single-use implementation
of a tree.

To fix this:

* define `TreeNode<T>`
* Switch `List<AbstractDeckTreeNode>` for either
  * `List<AbstractDeckTreeNode>` when a list was desired
  * `List<TreeNode<AbstractDeckTreeNode>>` when a tree was desired
* Fix all references to the structure
* Fix two creation patterns: filtering and creation in Sched
Filtering is easy - just filter the deck structures
 and create a new instance of the TreeNode with different children
 as all calculations were completed

Creation is less easy:
* define [processChildren] on `AbstractDeckTreeNode`
* call this with the children which will be added to the node