Improve performance for reviewer_*.xml layouts by using the merge tag
The reviewer_flashcard_*.xml layouts all have the same structure with a root FrameLayout and 3 full
screen FrameLayouts corresponding to different areas for the card reviewer screen.

The root layout is only used to set some Relativelayout positioning layout rules in AbstractFlashcardViewe.initLayout()
according to the user's button positioning choice. This root layout could be replaced by a merge tag and the work
done in AbstractFlashcardViewe.initLayout() on it could be replicated to each of its 3 children. This will
improve performance by making the view hierarchy shallower as the 3 children will be added directly to the
reviewer_*.xml layouts.