NF: Decoupling time from Collection
With experience, I came to agree that actually, time has nothing to do in the
collection itself. I don't like static values, but I guess it's still a better
solution. So here it is, Time is static. ("Time is static" is an absurd
sentence. I want to rename Time to Clock, but that'll wait)

I decided that when a new time object is created, it becomes the current
Time. Indeed, theer are no case I can think of where I want to create but not
use it. Nor case where I want to use two times simultaneously