NF: avoid !! on mMetaDb
It can't be lateinit because it sometime must legitimately be null.
While I don't specifically expect concurrency problem, it also has the nice
advantage that when a change is made on a database, all the changes of a
function are made. There is no risk for mMetaDB to become null in the middle of
a function.