Drop legacy Java backend implementation
As AnkiDroid begins to adopt more of the backend code, it will not be
practical to maintain a fallback for devices that are unable to load
the backend library. Assuming there are an acceptably few number of
DroidBackendFactory::getInstance crash reports, cutting the Java
fallback is one step towards a single concrete backend interface.

Also introduce AnkidroidApp.currentBackendFactory() to get the currently
active backend version, and switch a couple of calls that were using
BackendFactory.createInstance() (which hard-codes V11).