NF: waitForAllToFinish corrected
In 94aa0358e1da6a33f0f02519f4105db20dd2d249 I introduced an error.
The abstract `waitForAllToFinish` was deferring to `waitToFinishConcrete`
instead of `waitForAllToFinishConcrete`.

I have no knowledge of any issue it corrected, and I'm surprised by the
implementation of this method. Still, I correct it so that whether or not this
method is relevant can be discussed by itself.