Make the model property in ModelFieldEditor as not null
Although the mod(renamed to mModel) property can be null based on the type system, it being null doesn't make any sense
for this part of the code(currently it will crash the activity if this happens). I made this property as not null and handled the possible case a null model safely by showing a toast to the user and manually finishing the activity.