Implemented Notification Preference Datastore.
Notification preference datastore is used to store all the data related to
notification in the preference datastore.

We both introduce helper methods to deal with primitive type, and an ad-hoc type
to represents notification times for each decks in a way that is simple to store
in the datastore and efficient to process.