Adjust fab behavior to work with material snackbar
Previously, snackbars were sliding from the bottom. Material snackbars
instead fade in, and slightly scale out. These snackbars are meant to be
placed above the fab, but our fab opens a menu, and what are you going
to do with that, huh?

This makes fab slide up with some deceleration when a snackbar starts
appearing, and down after it is removed. This plays well with fab's
rotating animation.