Support backend undo
Can be tested by importing an apkg from the deck list, then choosing
the undo option.

- If a backend undo entry is present, it takes priority over the old
undo system to ensure data integrity. AD's reviewer saves config vars and
decks as part of the review process, and it was triggering undo entries
for those changes, which prevent the v2 scheduler review from being undone.
The calls have been updated to clear the backend undo, so that the review
takes precedence. The desktop code has the same problem, and it will be
avoidable once updating to the v3 scheduler.
- Introduce a helper to await a job in unit tests, to avoid race
conditions