Fix full download crashing in new schema case
This worked in my initial testing, but with a smaller collection, the
UI is now racing with the download on my machine:

- getSyncStatus() was assuming col.db would still be open by the time
it gets around to calling methods on it
- onCreateOptionsMenu() likewise assumes that if colIsOpen() returns true,
it is free to call methods on it

A better way way to fix this would be to wrap all collection access in a mutex,
so that the collection is never touched without the lock being acquired.
Code holding the mutex could then be guaranteed that the collection is open
and valid for the duration of its run. Changing all that legacy code is
a big task however.

I've added a lock to the upload case as well, as it's possible it's likewise
buggy and just not triggering on my machine.