Simplify setting snackbar behavior
Previously, snackbar behavior was set by by changing behavior of view's
CoordinatorLayout.LayoutParams. This came with a few downsides:

  * We had to set it in a Snackbar.Callback. Snackbar's original
    behavior is created when it is shown, besides, we needed the
    original listener.
  * It required a few casts.
  * We had to steal the original listener since it is useful for
    pausing the snackbars.

With this, behavior is set “the official way”, by calling
Snackbar.setBehavior. This comes with the only downside:

  * Snackbar.setBehavior expects a BaseTransientBottomBar.Behavior,
    which extends SwipeDismissBehavior, so we get it on the superclass
    list. It seems that this class wasn't meant to be subclassed.

While now we have a funny class on the parent list, setting behavior
“the official way” simplifies the code quite a bit, so this is probably
a worthwhile change.