Remove workaround in SwipeDismissBehaviorFix
This removes the workaround which prevented other views from being moved
while trying to move the snackbar. Instead, BaseTransientBottomBar
.Behavior logic is restored.

The removed workaround was applicable to SwipeDismissBehavior. The idea
was taken from an old StackOverflow featuring an old version of
Behavior. It worked by checking by only ever intercepting gestures if
the finger is currently on top of the snackbar.

BaseTransientBottomBar.Behavior, which Snackbar.setBehavior expects,
overrides canSwipeDismissView, and checks if the child is
SnackbarBaseLayout. canSwipeDismissView is called by the original
ViewDragHelper.Callback; since we have rolled out our own, it's best to
inline this check.