**Remove NumberRangePreferenceCompat onDialogClosed override
The only difference between NumberRangePreferenceCompat.onDialogClosed and super.onDialogClosed is that NumberRangePreferenceCompat's method doesn't call the preference change listener.

As some of the current preferences need a preference change listener and there isn't a reason to ignore it (nor it had one when the method was introduced), I'm removing the override.