fix: Set theme to Theme_Light for Layout Designer
Should only affect the Android Studio layout preview (designer).

If unset, the default is `Material.Light`, which does not
preview correctly (no blue menu bar at the top).

This has been tested and does not seem to have a functional effect on
the app.

If it does have a functional change, this change may be reverted

- the default theme for all activities, but we call `Themes.setTheme`
inside AnkiActivity, so this should only affect the layout designer.

https://developer.android.com/guide/topics/manifest/application-element#theme