Allow unsetting Sync url & Media sync url
When either of the settings is unset, AnkiWeb sync server is used. But
if you change any of these settings, you can't clear them.

This allows clearing the settings by inputting empty string.
This will save a string, not a null; however,
CustomSyncServer.getCollectionSyncUrlIfSetAndEnabledOrNull() will
return null for an empty string as well as for a null.