Do not hardcode /sync/ suffix in Sync url preference
Currently the preference Custom sync server -> Sync url must be
specified without the full endpoint path, e.g. http://foo:8080,
not http://foo:8080/sync/. The /sync/ path is hardcoded.
This is not optimal as:

* It's inconsistent with the preference Media sync url below;
* Anki uses the full URL in its --syncserver output, and expect the
  full URL in SYNC_ENDPOINT;
* Power users may want to use something other than /sync/ for security
  or other reasons.

This changes the preference to take arbitrary path, and migrates the
old preference.

In HttpSyncer.kt and RemoteMediaServer.kt I opted for creating
overridable methods getDefaultSyncUrl() and getCustomSyncUrlOrNull()
instead of putting everything into syncURL(). This is a tad un-pretty,
but it allows solving a slight issue with parseUrl() in a simple way.
parseUrl() used to called the now deleted isUsingCustomSyncServer()
to determine whether a custom URL was used; this was incorrect.
The latter would yield the state of the Use custom sync server switch
preference, however, even if the switch is on, the custom sync server--
collection or media--is not used if it is not set; instead, the default
AnkiWeb URL is used instead. This is confusing both in the code and
in the UI, and I hope it is eventually simplified.