[GSoC] Simplify add note task (#12070)
* Removed unnecessary callbacks

When adding a single note, it is not required to show continuous progress since it a short task and user can be notified after the task has been finished. So instead of using onProgressUpdate callback, it is good to use onPostExecute for UI updates while adding notes.

* Code simplification in NoteEditor.SaveNoteHandler

Both the conditions perform same task, sequential execution of both eventually nullifies the effect of !closeEditorAfterSave, so can be removed.