add insets to make the drawable icons center inside the checkbox
Insets is used to make the drawable icons which are of (24x24) to be in center of the checkbox which is (48x48). The value of insets is set only for left side as by default the drawable icon inside the checkbox is attached to the left side. So making the left inset value to 12dp moves the icon by 12dp to the right which makes it to go in the center of the checkbox automatically